% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef}
\alias{coef}
\title{Standard generic for "coef" defined from package "stats"}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{fitted model}

\item{...}{Additional arguments passed to other functions}
}
\description{
same as `coefficients()`
}
