% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_modsem_da.R
\name{coef_modsem_da}
\alias{coef_modsem_da}
\title{Wrapper for coef}
\usage{
coef_modsem_da(object, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for coef, to be used with modsem::coef_modsem_da, 
since coef is not in the namespace of modsem, but stats
}
