% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coefficients}
\alias{coefficients}
\title{Standard generic for "coefficients" defined from package "stats"}
\usage{
coefficients(object, ...)
}
\arguments{
\item{object}{fitted model}

\item{...}{Additional arguments passed to other functions}
}
\description{
same as `coef()`
}
