% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_output.R
\name{colorize_output}
\alias{colorize_output}
\title{Capture, colorise, and emit console text}
\usage{
colorize_output(
  expr,
  positive = MODSEM_COLORS$positive,
  negative = MODSEM_COLORS$negative,
  true = MODSEM_COLORS$true,
  false = MODSEM_COLORS$false,
  nan = MODSEM_COLORS$nan,
  na = MODSEM_COLORS$na,
  inf = MODSEM_COLORS$inf,
  string = MODSEM_COLORS$string,
  append = "\\n"
)
}
\arguments{
\item{expr}{Expression or object with output which should be colorized.}

\item{positive}{color of positive numbers.}

\item{negative}{color of negative numbers.}

\item{true}{color of \code{TRUE}.}

\item{false}{color of \code{FALSE}.}

\item{nan}{color of \code{NaN}.}

\item{na}{color of \code{NA}.}

\item{inf}{color of \code{-Inf} and \code{Inf}.}

\item{string}{color of quoted strings.}

\item{append}{String appended after the colored output (default `\\n`).}
}
\value{
Invisibly returns the *plain* captured text.
}
\description{
Capture, colorise, and emit console text
}
\examples{

set_modsem_colors(positive = "green", 
                  negative = "red",
                  true = "green", 
                  false = "red", 
                  na = "purple")

m1 <- "
# Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3
# Inner Model
  Y ~ X + Z + X:Z
"

est <- modsem(m1, data = oneInt)
colorize_output(summary(est))
colorize_output(print(est))

\dontrun{
est_lms <- modsem(m1, data = oneInt, method = "lms")

colorize_output(summary(est_lms))

colorize_output(modsem_inspect(est_lms))
}
}
