% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_settings_pi.R
\name{default_settings_pi}
\alias{default_settings_pi}
\title{default arguments for product indicator approaches}
\usage{
default_settings_pi(method = c("rca", "uca", "pind", "dblcent", "ca"))
}
\arguments{
\item{method}{which method to get the settings for}
}
\value{
modsem_lms or modsem_qml object
}
\description{
This function returns the default settings for the product indicator approaches
}
\examples{
library(modsem)
default_settings_pi()
}
