% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emLMS.R
\name{emLms}
\alias{emLms}
\title{Em Lms}
\usage{
emLms(
  model,
  verbose = FALSE,
  convergence = 0.01,
  maxiter = 500,
  maxstep = 1,
  negHessian = TRUE,
  breakOnLogLikIncrease = FALSE,
  ...
)
}
\arguments{
\item{model}{model object}

\item{...}{}
}
\description{
Em Lms
}
