% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_h0_da.R
\name{estimate_h0}
\alias{estimate_h0}
\title{Estimate baseline model for the LMS and QML approach}
\usage{
estimate_h0(object, ...)
}
\arguments{
\item{object}{An object of class `modsem_da`.}

\item{...}{Additional arguments passed to the `modsem_da` function, overriding
the arguments in the original model.}
}
\description{
Estimates a baseline model (H0) from a given model (H1) and compares the fit of both models.
The baseline model is estimated by removing all interaction terms from the model.
}
\examples{
\dontrun{
m1 <- "
 # Outer Model
 X =~ x1 + x2 + x3
 Y =~ y1 + y2 + y3
 Z =~ z1 + z2 + z3

 # Inner model
 Y ~ X + Z + X:Z
"

est_h1 <- modsem(m1, oneInt, "lms")
est_h0 <- estimate_h0(est_h1, calc.se=FALSE) # std.errors are not needed
compare_fit(est_h0, est_h1)
}
}
