% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsem_pi.R
\name{extract_lavaan}
\alias{extract_lavaan}
\title{extract lavaan object from modsem object estimated using product indicators}
\usage{
extract_lavaan(object)
}
\arguments{
\item{object}{modsem object}
}
\value{
lavaan object
}
\description{
extract lavaan object from modsem object estimated using product indicators
}
\examples{
library(modsem)
m1 <- '
  # Outer Model
  X =~ x1 + x2 + x3
  Y =~ y1 + y2 + y3
  Z =~ z1 + z2 + z3
  
  # Inner model
  Y ~ X + Z + X:Z 
'
est <- modsem_pi(m1, oneInt)
lav_est <- extract_lavaan(est) 
}
