% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lapplyDf}
\alias{lapplyDf}
\title{Wrapper of lapply for dataframes}
\usage{
lapplyDf(df, FUN, ...)
}
\arguments{
\item{df}{a dataframe}

\item{FUN}{a function to apply to variables in df}

\item{...}{}
}
\description{
Wrapper of lapply for dataframes
}
\examples{
lapplyDf(iris[1:4], FUN = sqrt)
}
