% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_coef}
\alias{modsem_coef}
\title{Wrapper for coef}
\usage{
modsem_coef(object, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for coef, to be used with modsem::modsem_coef, since
coef is not in the namespace of modsem, but stats
since coef is not in the namespace of modsem, but stats
}
