% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_nobs}
\alias{modsem_nobs}
\title{Wrapper for nobs}
\usage{
modsem_nobs(object, ...)
}
\arguments{
\item{object}{fitted model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for \code{nobs}, to be used with \code{modsem::modsem_nobs}, since
\code{nobs} is not in the namespace of \code{modsem}, but \code{stats}.
}
