% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_vcov}
\alias{modsem_vcov}
\title{Wrapper for vcov}
\usage{
modsem_vcov(object, ...)
}
\arguments{
\item{object}{fitted model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for \code{vcov}, to be used with \code{modsem::modsem_vcov}, since
\code{vcov} is not in the namespace of \code{modsem}, but \code{stats}.
}
