% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiplyIndicatorsCpp}
\alias{multiplyIndicatorsCpp}
\title{Multiply indicators}
\usage{
multiplyIndicatorsCpp(df)
}
\arguments{
\item{df}{A data DataFrame}
}
\value{
A NumericVector
}
\description{
Multiply indicators
}
