% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nobs}
\alias{nobs}
\title{Get number of observations}
\usage{
nobs(object, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{...}{additional arguments}
}
\description{
Get number of observations
}
