% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{parameter_estimates}
\alias{parameter_estimates}
\title{Extract parameterEstimates from an estimated model}
\usage{
parameter_estimates(object, ...)
}
\arguments{
\item{object}{An object of class `modsem_pi`, `modsem_lms_qml`, or `modsem_mplus`}

\item{...}{Additional arguments passed to other functions}
}
\description{
Extract parameterEstimates from an estimated model
}
