% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelSpecification.R
\name{parseLavaan}
\alias{parseLavaan}
\title{Parse lavaan model}
\usage{
parseLavaan(modelSyntax = NULL, variableNames = NULL, match = FALSE)
}
\arguments{
\item{modelSyntax}{lavaan syntax}

\item{isMeasureSpecified}{have you specified the measure model for the latent prod}
}
\description{
Parse lavaan model
}
