% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLMS.R
\name{print.summaryModsemLMS}
\alias{print.summaryModsemLMS}
\title{print lms}
\usage{
\method{print}{summaryModsemLMS}(x, digits = 3, ...)
}
\arguments{
\item{x}{}

\item{digits}{}

\item{...}{}
}
\description{
print lms
}
