% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{standardized_estimates}
\alias{standardized_estimates}
\title{Get standardized estimates}
\usage{
standardized_estimates(object, ...)
}
\arguments{
\item{object}{An object of class `modsem_lms_qml`,  `modsem_mplus`, 
or a parTable of class `data.frame`}

\item{...}{Additional arguments passed to other functions}
}
\description{
Get standardized estimates
}
