% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_lms_qml.R, R/generics_modsem_mplus.R,
%   R/generics_modsem_pi.R
\name{summary.modsem_lms}
\alias{summary.modsem_lms}
\alias{summary.modsem_qml}
\alias{summary.modsem_mplus}
\alias{summary.modsem_pi}
\title{summary.modsem_lms}
\usage{
\method{summary}{modsem_lms}(object, ...)

\method{summary}{modsem_qml}(object, ...)

\method{summary}{modsem_mplus}(object, ...)

\method{summary}{modsem_pi}(object, ...)
}
\arguments{
\item{object}{modsem object to summarized}

\item{...}{arguments passed to lavaan::summary()}
}
\description{
summary.modsem_lms

summary.modsem_qml

summary.modsem_mplus

summary.modsem_pi
}
