% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsem.R
\name{summary,ModSEM-method}
\alias{summary,ModSEM-method}
\title{summary.ModSEM}
\usage{
\S4method{summary}{ModSEM}(object, ...)
}
\arguments{
\item{object}{modsem object to summarized}
}
\description{
summary.ModSEM
}
