% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathTracer.R
\name{tracePath}
\alias{tracePath}
\title{Estimate formulas for (co-)variance paths using Wright's path tracing rules}
\usage{
tracePath(pt, x, y, ...)
}
\arguments{
\item{pt}{A data frame with columns lhs, op, rhs, and mod, from modsemify(syntax)}

\item{x}{source variable}

\item{y}{destination variable}

\item{...}{additional arguments passed to tracePath}
}
\value{
A string with the estimated path (simplified if possible)
}
\description{
This function estimates the path from x to y using the path tracing rules, 
note that it only works with structural parameters, so "=~" are ignored. If you
you want to use the measurement model, it should work if you replace it "=~" with
"~" in the mod column of pt.
}
