% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{var_interactions}
\alias{var_interactions}
\title{Extract or modify parTable from an estimated model with estimated variances of interaction terms}
\usage{
var_interactions(object, ...)
}
\arguments{
\item{object}{An object of class `modsem_da`,  `modsem_mplus`, 
or a parTable of class `data.frame`}

\item{...}{Additional arguments passed to other functions}
}
\description{
Extract or modify parTable from an estimated model with estimated variances of interaction terms
}
