% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov}
\alias{vcov}
\title{extract variance covariance matrix of model parameters}
\usage{
vcov(object, ...)
}
\arguments{
\item{object}{fitted model}

\item{...}{Additional arguments passed to other functions}
}
\description{
extract variance covariance matrix of model parameters
}
