% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_modsem_da.R
\name{vcov_modsem_da}
\alias{vcov_modsem_da}
\title{Wrapper for vcov}
\usage{
vcov_modsem_da(object, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for vcov, to be used with modsem::vcov_modsem_da, 
since vcov is not in the namespace of modsem, but stats
}
