% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_model.R
\name{compile_stan_model}
\alias{compile_stan_model}
\title{Compile \code{STAN} model based on a \code{lavaan} model}
\usage{
compile_stan_model(model.syntax, compile = TRUE)
}
\arguments{
\item{model.syntax}{\code{lavaan} syntax.}

\item{compile}{Should compilation be performed? If \code{FALSE} only the \code{STAN}
is generated, and not compiled.}
}
\description{
Compile \code{STAN} model based on a \code{lavaan} model
}
