% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_settings_da.R
\name{default_settings_da}
\alias{default_settings_da}
\title{default arguments fro LMS and QML approach}
\usage{
default_settings_da(method = c("lms", "qml"))
}
\arguments{
\item{method}{which method to get the settings for}
}
\value{
list
}
\description{
This function returns the default settings for the LMS and QML approach.
}
\examples{
library(modsem)
default_settings_da()
}
