% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsem2.1.R
\name{modsem}
\alias{modsem}
\title{Interaction between latent variables}
\usage{
modsem(
  modelSyntax = NULL,
  data = NULL,
  method = "rca",
  standardizeData = FALSE,
  isMeasurementSpecified = FALSE,
  firstLoadingFixed = TRUE,
  doubleCentering = FALSE,
  centeredProducts = FALSE,
  residualsProducts = FALSE,
  residualCovSyntax = FALSE,
  constrainedProductMean = FALSE,
  ...
)
}
\arguments{
\item{modelSyntax}{lavaan syntax}

\item{data}{dataframe}

\item{method}{method to use}

\item{standardizeData}{should data be scaled before fitting model}

\item{isMeasurementSpecified}{have you specified the measurement model for the latent product}

\item{firstLoadingFixed}{Sould the first factorloading in the latent product be fixed to one?}

\item{doubleCentering}{should indicatorproducts be centered after they have been computed?}

\item{centeredProducts}{should indicators in products be centered (overwritten by method, if method != NULL)}

\item{residualsProducts}{should indicator products be centered using residuals (overwritten by method, if method != NULL)}

\item{residualCovSyntax}{should syntax for residual covariances be produced (overwritten by method, if method != NULL)}

\item{constrainedProductMean}{should syntax product mean be produced (overwritten by method, if method != NULL)}

\item{...}{arguments passed to other functions, e.g,. lavaan}
}
\description{
Interaction between latent variables
}
