% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_coef}
\alias{modsem_coef}
\title{Wrapper for coef}
\usage{
modsem_coef(object, ...)
}
\arguments{
\item{object}{fitted model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for \code{coef}, to be used with \code{modsem::modsem_coef}, since
\code{coef} is not in the namespace of \code{modsem}, but \code{stats}.
}
