% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_inspect}
\alias{modsem_inspect}
\title{Inspect model information}
\usage{
modsem_inspect(object, what = NULL, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{what}{what to inspect}

\item{...}{Additional arguments passed to other functions}
}
\description{
function used to inspect fittet object. similar to `lavInspect()`
argument 'what' decides what to inspect
}
\details{
for `modsem_da`, and `modsem_lavaan`
for `modsem_lavaan`, it is just a wrapper for `lavInspect()`
for `modsem_da` and `` what can either be "all", "matrices", "optim",
or just the name of what to extract.
}
