% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsem_mplus.R
\name{modsem_mplus}
\alias{modsem_mplus}
\title{Estimation latent interactions through \code{Mplus}}
\usage{
modsem_mplus(
  model.syntax,
  data,
  estimator = "ml",
  type = "random",
  algorithm = "integration",
  process = 8,
  integration = 15,
  ...
)
}
\arguments{
\item{model.syntax}{lavaan/modsem syntax}

\item{data}{dataset}

\item{estimator}{estimator argument passed to \code{Mplus}}

\item{type}{type argument passed to \code{Mplus}}

\item{algorithm}{algorithm argument passed to \code{Mplus}}

\item{process}{process argument passed to \code{Mplus}}

\item{integration}{integration argument passed to \code{Mplus}}

\item{...}{arguments passed to other functions}
}
\value{
modsem_mplus object
}
\description{
Estimation latent interactions through \code{Mplus}
}
\examples{
# Theory Of Planned Behavior
m1 <- '
# Outer Model
  X =~ x1 + x2
  Z =~ z1 + z2
  Y =~ y1 + y2 

# Inner model
  Y ~ X + Z + X:Z
'

\dontrun{
# Check if Mplus is installed
run <- tryCatch({MplusAutomation::detectMplus(); TRUE},
                error = \(e) FALSE)

if (run) {
  est_mplus <- modsem_mplus(m1, data = oneInt)
  summary(est_mplus)
}
}

}
