% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_nobs}
\alias{modsem_nobs}
\title{Wrapper for nobs}
\usage{
modsem_nobs(object, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for nobs, to be used with modsem::modsem_nobs, since
nobs is not in the namespace of modsem, but stats
since nobs is not in the namespace of modsem, but stats
}
