% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modsem_vcov}
\alias{modsem_vcov}
\title{Wrapper for vcov}
\usage{
modsem_vcov(object, ...)
}
\arguments{
\item{object}{fittet model to inspect}

\item{...}{additional arguments}
}
\description{
wrapper for vcov, to be used with modsem::modsem_vcov, since
vcov is not in the namespace of modsem, but stats
since vcov is not in the namespace of modsem, but stats
}
