% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{modsemify}
\alias{modsemify}
\title{Generate parameter table for lavaan syntax}
\usage{
modsemify(syntax)
}
\arguments{
\item{syntax}{}
}
\value{
data.frame with columns lhs, op, rhs, mod
}
\description{
Generate parameter table for lavaan syntax
}
