% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{parameter_estimates}
\alias{parameter_estimates}
\title{Extract parameterEstimates from an estimated model}
\usage{
parameter_estimates(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{modsem_pi}}, \code{\link{modsem_da}}, or \code{\link{modsem_mplus}}}

\item{...}{Additional arguments passed to other functions}
}
\description{
Extract parameterEstimates from an estimated model
}
\examples{
m1 <- '
  # Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3

  # Inner Model
  Y ~ X + Z + X:Z
'
# Double centering approach
est_dca <- modsem(m1, oneInt)

pars <- parameter_estimates(est_dca) # no correction

# Pretty summary
summarize_partable(pars)

# Only print the data.frame
pars
}
