% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLavaan_v2_1.R
\name{parseLavaan}
\alias{parseLavaan}
\title{Parse lavaan model}
\usage{
parseLavaan(modelSyntax = NULL, isMeasurementSpecified = FALSE)
}
\arguments{
\item{modelSyntax}{lavaan syntax}

\item{isMeasurementSpecified}{have you specified the measurement model for the latent product}
}
\description{
Parse lavaan model
}
