% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Plot Interaction Effects}
\usage{
plot_interaction(
  x,
  z,
  y,
  xz,
  vals_x = seq(-3, 3, 0.001),
  vals_z,
  model,
  alpha_se = 0.15,
  ...
)
}
\arguments{
\item{x}{The name of the variable on the x-axis}

\item{z}{The name of the moderator variable}

\item{y}{The name of the outcome variable}

\item{xz}{The name of the interaction term}

\item{vals_x}{The values of the x variable to plot, the more values the smoother the std.error-area will be}

\item{vals_z}{The values of the moderator variable to plot. A seperate regression 
line ("y ~ x | z") will be plotted for each value of the moderator variable}

\item{model}{An object of class `modsem_pi`, `modsem_lms_qml`, or `modsem_mplus`}

\item{alpha_se}{The alpha level for the std.error area}

\item{...}{Additional arguments passed to other functions}
}
\value{
A ggplot object
}
\description{
Plot Interaction Effects
}
