% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_modsem_mplus.R,
%   R/generics_modsem_pi.R, R/modsem_pi.R
\name{summary.modsem_mplus}
\alias{summary.modsem_mplus}
\alias{summary.modsem_pi}
\alias{summary.ModSEM}
\title{summary.modsem_mplus}
\usage{
\method{summary}{modsem_mplus}(object, ...)

\method{summary}{modsem_pi}(object, ...)

\method{summary}{ModSEM}(object, ...)
}
\arguments{
\item{object}{modsem object to summarized}

\item{...}{arguments passed to lavaan::summary(), and nlsem::summary()}
}
\description{
summary.modsem_mplus

summary.modsem_pi

summary.ModSEM
}
