% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsem.R
\name{summary.ModSEM}
\alias{summary.ModSEM}
\title{summary.ModSEM}
\usage{
\method{summary}{ModSEM}(object, ...)
}
\arguments{
\item{object}{modsem object to summarized}

\item{...}{arguments passed to lavaan::summary(), and nlsem::summary()}
}
\description{
summary.ModSEM
}
