% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_lms_qml.R
\name{summaryLmsAndQml}
\alias{summaryLmsAndQml}
\title{summaryLmsAndQml}
\usage{
summaryLmsAndQml(
  object,
  H0 = TRUE,
  verbose = TRUE,
  r.squared = TRUE,
  digits = 3,
  scientific = FALSE,
  ci = FALSE,
  ...
)
}
\arguments{
\item{object}{modsem object estimated with lms or qml}

\item{H0}{should a null model be estimated (used for comparison)}

\item{verbose}{print progress for the estimation of null model}

\item{r.squared}{calculate R-squared}

\item{digits}{number of digits to print}

\item{scientific}{print p-values in scientific notation}

\item{ci}{print confidence intervals}

\item{...}{additional arguments}
}
\description{
summaryLmsAndQml
}
\examples{
\dontrun{
m1 <- '
 # Outer Model
 X =~ x1 + x2 + x3
 Y =~ y1 + y2 + y3
 Z =~ z1 + z2 + z3
 
 # Inner model
 Y ~ X + Z + X:Z 
'

est1 <- modsem(m1, oneInt, "qml")
summary(est1, ci = TRUE, scientific = TRUE)
}
}
